# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Laravel",
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "http://localhost/api"
        }
    ],
    "paths": {
...
        "/participants/{id}/transcript": {
            "get": {
                "operationId": "participant.transcript",
                "description": "Transcript scope (D2) \u2014 requires lifecycle >= in_valutazione; a\npre-threshold status raises LifecycleNotReadyException -> 409 (D4).",
                "summary": "GET /api/participants/{id}/transcript",
                "tags": [
                    "Participant"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`TranscriptResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/TranscriptResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
        "/participants/{id}/evaluation": {
            "get": {
                "operationId": "participant.evaluation",
                "description": "Evaluation scope (D2) \u2014 requires lifecycle === completato; anything\nelse raises LifecycleNotReadyException -> 409 (D4).",
                "summary": "GET /api/participants/{id}/evaluation",
                "tags": [
                    "Participant"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`EvaluationResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/EvaluationResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "EvaluationResource": {
                "type": "object",
                "additionalProperties": {},
                "title": "EvaluationResource"
            },
...
            "TranscriptResource": {
                "type": "object",
                "properties": {
                    "sessions": {
                        "type": "string"
                    }
                },
                "required": [
                    "sessions"
                ],
                "title": "TranscriptResource"
            },
...
        },
        "responses": {
...
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
...
        }
    }
}